Implement searchable question listing
